chore: add strict Python quality baseline (ruff + mypy strict)
Adds ruff (extended ruleset) and mypy strict=true to the bronze gate.
All existing Python files brought to zero violations: type annotations
throughout, stubs for numba/sklearn/sklearn.decomposition, no #noqa,
no inline # type: ignore. Per-file-ignores document legitimate domain
conventions (uppercase sklearn matrices, lazy training imports).
Unsafe-fixes workflow documented in CLAUDE.md.

+++ killer_sudoku/training/calibrate.py
@@ -90,7 +90,7 @@ def calibrate_isblack_offset(puzzle_dir: Path, n_images: int = 20) -> int:
     if not diffs:
         raise ValueError("No images could be read from puzzle_dir")
 
-    recommended = int(round(float(np.mean(diffs))))
+    recommended = round(float(np.mean(diffs)))
     _log.info(
         "Calibrated isblack_offset=%d from %d images (mean diff=%.1f)",
         recommended,
